memory: go tickless

+++ src/memory.cc
@@ -15,7 +15,7 @@ void Memory::packets_received( const vector< Packet > & packets, const unsigned 
       continue;
     }
 
-    const unsigned int rtt = x.tick_received - x.tick_sent;
+    const double rtt = x.tick_received - x.tick_sent;
     if ( _last_tick_sent == 0 || _last_tick_received == 0 ) {
       _last_tick_sent = x.tick_sent;
       _last_tick_received = x.tick_received;
